+++ packages/terra-dropdown-button/tests/wdio/split-button-spec.js
@@ -25,9 +25,23 @@ Terra.describeViewports('Split Button', ['medium'], () => {
 
     it('should validate primary button active styling', () => {
       $('[class*=split-button-primary]').moveTo();
-      browser.buttonDown();
+      browser.performActions([{
+        'type': 'pointer',
+        'id': 'pointer1',
+        'parameters': {'pointerType': 'mouse'},
+        'actions': [
+            {'type': 'pointerDown', 'button': 0}
+        ]
+      }]);
       Terra.validates.element('primary active');
-      browser.buttonUp();
+      browser.performActions([{
+        'type': 'pointer',
+        'id': 'pointer2',
+        'parameters': {'pointerType': 'mouse'},
+        'actions': [
+            {'type': 'pointerUp', 'button': 0}
+        ]
+      }]);
     });
 
     it('opens and validates dropdown with click', () => {
@@ -181,8 +195,15 @@ Terra.describeViewports('Split Button', ['medium'], () => {
 
       it('closes the dropdown when clicking outside the dropdownbutton', () => {
         $('#root').moveTo({ xOffset: 200, yOffset: 1 });
-        browser.buttonDown();
-        browser.buttonUp();
+        browser.performActions([{
+          'type': 'pointer',
+          'id': 'pointer3',
+          'parameters': {'pointerType': 'mouse'},
+          'actions': [
+              {'type': 'pointerDown', 'button': 0},
+              {'type': 'pointerUp', 'button': 0}
+          ]
+        }]);
         Terra.validates.element('clicking outside closes dropdown');
       });