+++ packages/terra-signature/tests/wdio/signature-spec.js
@@ -9,28 +9,52 @@ Terra.describeViewports('Signature', ['medium'], () => {
 
   it('should try to draw with right click', () => {
     $('#drawline').moveTo({ xOffset: 0, yOffset: 0 });
-    browser.buttonDown(2);
-    $('#drawline').moveTo({ xOffset: 90, yOffset: 90 });
-    browser.buttonUp(2);
+    browser.performActions([{
+      'type': 'pointer',
+      'id': 'pointer1',
+      'parameters': {'pointerType': 'mouse'},
+      'actions': [
+          {'type': 'pointerDown', 'button': 2},
+          {"type": "pointerMove", "duration": 0, "x": 90, "y": 90},
+          {'type': 'pointerUp', 'button':2}
+      ]
+   }]);
 
     Terra.validates.element('draw with right click');
+    browser.releaseActions();
   });
 
   it('should try to draw with middle click', () => {
     $('#drawline').moveTo({ xOffset: 0, yOffset: 0 });
-    browser.buttonDown(1);
-    $('#drawline').moveTo({ xOffset: 90, yOffset: 90 });
-    browser.buttonUp(1);
+    browser.performActions([{
+      'type': 'pointer',
+      'id': 'pointer2',
+      'parameters': {'pointerType': 'mouse'},
+      'actions': [
+          {'type': 'pointerDown', 'button': 1},
+          {"type": "pointerMove", "duration": 0, "x": 90, "y": 90},
+          {'type': 'pointerUp', 'button': 1}
+      ]
+   }]);
 
     Terra.validates.element('draw with middle click');
+    browser.releaseActions();
   });
 
   it('should draws lines', () => {
     $('#drawline').moveTo({ xOffset: 0, yOffset: 0 });
-    browser.buttonDown(0);
-    $('#drawline').moveTo({ xOffset: 90, yOffset: 90 });
-    browser.buttonUp(0);
+    browser.performActions([{
+      'type': 'pointer',
+      'id': 'pointer3',
+      'parameters': {'pointerType': 'mouse'},
+      'actions': [
+          {'type': 'pointerDown', 'button': 0},
+          {"type": "pointerMove", "duration": 0, "x": 90, "y": 90},
+          {'type': 'pointerUp', 'button': 0}
+      ]
+   }]);
 
     Terra.validates.element('draws lines');
+    browser.releaseActions();
   });
 });